Agregando los cambios a la versión anterior al refactor

+++ grails-app/services/com/payable/ComprobanteService.groovy
@@ -34,9 +34,36 @@ class ComprobanteService {
     pago
   }
 
+  def aprobarPago(String transactionId, Date fechaDePago, def tipoPago,String referencia){
+    def pago = Pago.findByTransactionId(transactionId)
+    pago.tipoDePago = TipoDePago.getAt(tipoPago)
+    pago.fechaDePago = fechaDePago
+    pago.referencia = referencia
+    pago.estatusDePago = (pago.tipoDePago == TipoDePago.EFECTIVO ? EstatusDePago.PAGADO : EstatusDePago.PROCESO)
+    pago.descuentosAplicables.findAll { da ->
+      da.descuentoAplicableStatus = DescuentoAplicableStatus.VIGENTE
+    }*.descuentoAplicableStatus = DescuentoAplicableStatus.APLICADO
+    pago.save()
+    pago
+  }
+
+  def aprobarPagoConciliacion(String transactionId, Date fechaDePago, def tipoPago, String referencia){
+    def pago = Pago.findByTransactionId(transactionId)
+    pago.tipoDePago = TipoDePago.getAt(tipoPago)
+    pago.fechaDePago = fechaDePago
+    pago.referencia = referencia
+    pago.estatusDePago = EstatusDePago.PAGADO
+    pago.descuentosAplicables.findAll { da ->
+      da.descuentoAplicableStatus = DescuentoAplicableStatus.VIGENTE 
+    }*.descuentoAplicableStatus = DescuentoAplicableStatus.APLICADO
+    pago.save()
+    pago
+  }
+
   def rechazarPago(String transactionId) {
     def pago = Pago.findByTransactionId(transactionId)
-    s3AssetService.delete(pago.comprobanteDePago)
+    if(pago.comprobanteDePago)
+      s3AssetService.delete(pago.comprobanteDePago)
     pago.comprobanteDePago = null
     pago.estatusDePago = EstatusDePago.RECHAZADO
     pago.save()
